made changes and split some things up
