+++ tutorials/3.Interactions.md
@@ -181,8 +181,8 @@ As before, we'll start by moving the helper function (`pingDB`) to its own top l
 Doing this though, we immediately come up on the snag. Even if we aren't sending a message using `m` any more, we are still getting the user id of the user that sent the message. To solve this, we have to perform a special kind of parsing, using the context we have. This changes the first couple lines of `pingDB` to the following.
 
 ```haskell
-pingDB :: UserId -> DatabaseDiscord MessageDetails
-pingDB u = do
+pingDB :: SenderUserId -> DatabaseDiscord MessageDetails
+pingDB (SenderUserId u) = do
     let uid = extractFromSnowflake $ unId u
 ```
 
@@ -315,8 +315,8 @@ myEchoInteraction = makeApplicationCommandPair "myecho" "echo your input" (echoH
 myPing'' :: Command
 myPing'' = Command "cmyping" (parseComm pingDB) []
 
-pingDB :: UserId -> DatabaseDiscord MessageDetails
-pingDB u = do
+pingDB :: SenderUserId -> DatabaseDiscord MessageDetails
+pingDB (SenderUserId u) = do
     let uid = extractFromSnowflake $ unId u
     user <- liftSql $ select $ from $ \p -> do
         where_ (p ^. PingCountUid ==. val uid)
